changed Subscriber to use generics
