+++ toy-rpc/src/server/pubsub.rs
@@ -94,3 +94,4 @@ impl PubSubBroker {
         }
     }
 }
+
